libs/requirements.txt
marshmallow==3.2.2
numpy==1.17.4
pandas==0.25.3
pandas-schema==0.3.4
python-dateutil==2.8.1

pytest
pytest-cov
pytest-runner
Faker
coverage
coverage-badge==0.2.0
Fabric==2.4.0
snapshottest
factory_boy
mongomock
flake8
